Put `GpuBufferFormatExtension` in its own file
