+++ Mediapipe.Net/Gpu/GpuBufferFormat.cs
@@ -2,9 +2,6 @@
 // This file is part of MediaPipe.NET.
 // MediaPipe.NET is licensed under the MIT License. See LICENSE for details.
 
-using Mediapipe.Net.Framework.Format;
-using Mediapipe.Net.Native;
-
 namespace Mediapipe.Net.Gpu
 {
     public enum GpuBufferFormat : uint
@@ -21,19 +18,5 @@ namespace Mediapipe.Net.Gpu
         kRgb24 = 0x00000018,  // Note: prefer Bgra32 whenever possible.
         kRgbaHalf64 = ('R' << 24) + ('G' << 16) + ('h' << 8) + 'A',
         kRgbaFloat128 = ('R' << 24) + ('G' << 16) + ('f' << 8) + 'A',
-    }
-
-    public static class GpuBufferFormatExtension
-    {
-        public static ImageFormat ImageFormatFor(this GpuBufferFormat gpuBufferFormat)
-        {
-            return SafeNativeMethods.mp__ImageFormatForGpuBufferFormat__ui(gpuBufferFormat);
-        }
-
-        public static GlTextureInfo GlTextureInfoFor(this GpuBufferFormat gpuBufferFormat, int plane, GlVersion glVersion = GlVersion.KGles3)
-        {
-            UnsafeNativeMethods.mp__GlTextureInfoForGpuBufferFormat__ui_i_ui(gpuBufferFormat, plane, glVersion, out var glTextureInfo).Assert();
-            return glTextureInfo;
-        }
     }
 }
